improve the deploy script

+++ deploy.sh
@@ -1,14 +1,27 @@
 #!/bin/bash
 
-TMPDIR=/tmp/adbpush/pytt
+# usage: deploy.sh [--all]
 
+TMPDIR=/tmp/adbpush/pytt
+FILES=(
+    android.txt
+    main.py
+    pytt
+    )
+
+if [ x$1 = "x--all" ]
+then
+    FILES+=(libs)
+    adb shell rm -rf /sdcard/kivy/pytabletop
+fi
+
+rm -rf $TMPDIR
 mkdir -p $TMPDIR
-cp -L -r * $TMPDIR
+cp -L -r "${FILES[@]}" $TMPDIR
 cd $TMPDIR
 py.cleanup
 cd -
 
-adb shell rm -rf /sdcard/kivy/pytabletop
 adb push $TMPDIR /sdcard/kivy/pytabletop
 adb logcat -c
 adb logcat -s python
